distributed_discrete_gaussian/requirements.txt
absl-py
attrs
matplotlib
numpy
pandas
scipy
tensorflow
tensorflow-federated
tensorflow-privacy
tensorflow-probability
